pyproject.toml
[project]
name = "gunlinuxblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "loki", email = "gunlinux@ya.ru" }
]
requires-python = ">=3.10"
dependencies = [
    "Flask-Admin>=1.6.1",
    "flask-caching>=2.3.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sitemap>=0.4.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gevent>=25.5.1",
    "gunicorn>=23.0.0",
    "markdown>=3.7",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "wtforms<3.0",
]

[project.scripts]
gunlinuxblog = "gunlinuxblog:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gunlinuxblog"]

[tool.pyright]
ignore = ["tests", "migrations"]
reportOptionalCall = false
reportUnusedCallResult = false
# enable?
reportUnannotatedClassAttribute = false
reportAny = false
reportMissingTypeStubs = false
reportUninitializedInstanceVariable = false
reportImportCycles = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportAttributeAccessIssue = false

[tool.coverage.run]
branch = true

[tool.pytest.ini_options]
norecursedirs = "__pycache__"
addopts = "--strict-markers --cov=blog --cov-report xml"
filterwarnings = [
    "ignore::UserWarning:.*pkg_resources.*",
    "ignore::DeprecationWarning:.*pkg_resources.*",
    "ignore::UserWarning:.*flask_admin.*",
    "ignore::DeprecationWarning:.*flask_admin.*",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.5",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]
